- enhancements to JavaParser - added some more tokens to Fortran - improvements with look ahead function of Java RegExp for better recognition   of keywords
git-svn-id: http://server1:10080/svn/PureSolTechnologiesAPI@92 695a9040-d605-4429-820e-8508b0f4e340

+++ trunk/src/com/puresol/coding/lang/java/source/keywords/LongKeyword.java
@@ -4,9 +4,9 @@ import com.puresol.coding.tokentypes.PrimitiveDataType;
 
 public class LongKeyword extends PrimitiveDataType {
 
-	@Override
-	public String getPatternString() {
-		return "short";
-	}
+    @Override
+    public String getPatternString() {
+	return "long";
+    }
 
 }
